Ingredient RecyclerView major refactor.
